Split io runtime from C driver
